+++ agrichat-backend/Agentic_RAG_/main.py
@@ -1,14 +1,15 @@
 from crewai import Crew
-from crew_agents import (
+from .crew_agents import (
     Retriever_Agent, Grader_agent,
     hallucination_grader, answer_grader
 )
-from crew_tasks import (
+from .crew_tasks import (
     retriever_task, grader_task,
     hallucination_task, answer_task
 )
 
 def get_answer(question):
+    print("here")
     rag_crew = Crew(
         agents=[
             Retriever_Agent
@@ -20,7 +21,17 @@ def get_answer(question):
     )
     inputs = {"question": question}
     result = rag_crew.kickoff(inputs=inputs)
-    return result
+    print(question)
+    print("Result object from crew.kickoff():", result)
+
+    if hasattr(result, "output"):
+        return result.output
+    elif hasattr(result, "text"):
+        return result.text
+    elif isinstance(result, str):
+        return result
+    else:
+        return str(result) 
 
 if __name__ == "__main__":
     while True:
