+++ agrichat-frontend/script.js
@@ -1,5 +1,5 @@
-const API_BASE = "https://agrichat-annam.onrender.com/api";
-// const API_BASE = "http://localhost:8000/api";
+// const API_BASE = "https://agrichat-annam.onrender.com/api";
+const API_BASE = "http://localhost:8000/api";
 
 const stateLanguageMap = {
   "Andhra Pradesh": "Telugu",
